fix: address PR review feedback for agent skills docs
- Fix error handling to throw on fetch failures instead of silently returning empty data
- Merge AiSkillsTable into AiSkillsIndex, extract CopyButton to its own client component
- Add TODO comment on breadcrumbs patch
